requirements.txt
aiogram==3.6.0
aiohttp==3.9.1
Flask==3.0.3
flask_sqlalchemy==3.1.1
Pillow==9.5.0
waitress==3.0.0
